allow bigger number in collatz and remove junk file

+++ collatz.c
@@ -2,15 +2,15 @@
 
 int main(int argc, char *argv[])
 {
-   int number;
-   scanf("%d", &number);
+   long number;
+   scanf("%ld", &number);
    
    while (number > 1) {
      if (number % 2 == 0) {
        number = number / 2;
      } else
        number = number * 3 + 1;
-     printf("%d\n", number);
+     printf("%ld\n", number);
    }
    
    return 0;
